+++ packages/tendermint-rpc/src/legacy/adaptors/v0-33/responses.ts
@@ -3,7 +3,9 @@ import { fromBase64, fromHex } from "@cosmjs/encoding";
 import { JsonRpcSuccessResponse } from "@cosmjs/json-rpc";
 import { assert } from "@cosmjs/utils";
 
-import { fromRfc3339WithNanoseconds } from "../../dates";
+import { fromRfc3339WithNanoseconds } from "../../../dates";
+import { SubscriptionEvent } from "../../../rpcclients";
+import { BlockIdFlag, CommitSignature, ValidatorPubkey } from "../../../types";
 import {
   assertArray,
   assertBoolean,
@@ -17,10 +19,8 @@ import {
   may,
   optional,
 } from "../../encodings";
+import { hashTx } from "../../hasher";
 import * as responses from "../../responses";
-import { SubscriptionEvent } from "../../rpcclients";
-import { BlockIdFlag, CommitSignature, ValidatorPubkey } from "../../types";
-import { hashTx } from "./hasher";
 
 interface AbciInfoResult {
   readonly response: RpcAbciInfoResponse;