+++ packages/tendermint-rpc/src/legacy/responses.ts
@@ -1,7 +1,7 @@
 import { ReadonlyDate } from "readonly-date";
 
-import { ReadonlyDateWithNanoseconds } from "./dates";
-import { CommitSignature, ValidatorPubkey } from "./types";
+import { ReadonlyDateWithNanoseconds } from "../dates";
+import { CommitSignature, ValidatorPubkey } from "../types";
 
 export type Response =
   | AbciInfoResponse