+++ packages/tendermint-rpc/src/tendermint34/adaptors/index.ts
@@ -1,0 +1,2 @@
+/* eslint-disable @typescript-eslint/naming-convention */
+export { v0_34 as adaptor34 } from "./v0-34";