+++ packages/tendermint-rpc/src/tendermint34/adaptors/v0-34/index.ts
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { Adaptor } from "../../adaptor";
-import { hashBlock, hashTx } from "./hasher";
+import { hashBlock, hashTx } from "../../hasher";
 import { Params } from "./requests";
 import { Responses } from "./responses";
 
-export const v0_33: Adaptor = {
+export const v0_34: Adaptor = {
   params: Params,
   responses: Responses,
   hashTx: hashTx,