+++ packages/tendermint-rpc/src/tendermint34/tendermint34client.spec.ts
@@ -5,46 +5,27 @@ import { sleep } from "@cosmjs/utils";
 import { ReadonlyDate } from "readonly-date";
 import { Stream } from "xstream";
 
-import { Adaptor } from "./adaptor";
-import { adaptorForVersion } from "./adaptors";
-import { Client } from "./client";
-import { ExpectedValues, tendermintInstances } from "./config.spec";
+import { HttpClient, RpcClient, WebsocketClient } from "../rpcclients";
+import {
+  buildKvTx,
+  chainIdMatcher,
+  ExpectedValues,
+  pendingWithoutTendermint,
+  randomString,
+  tendermintEnabled,
+  tendermintInstances,
+  tendermintSearchIndexUpdated,
+} from "../testutil.spec";
+import { adaptor34 } from "./adaptors";
 import { buildQuery } from "./requests";
 import * as responses from "./responses";
-import { HttpClient, RpcClient, WebsocketClient } from "./rpcclients";
-import { chainIdMatcher } from "./testutil.spec";
+import { Tendermint34Client } from "./tendermint34client";
 
-function tendermintEnabled(): boolean {
-  return !!process.env.TENDERMINT_ENABLED;
-}
-
-function pendingWithoutTendermint(): void {
-  if (!tendermintEnabled()) {
-    pending("Set TENDERMINT_ENABLED to enable tendermint-based tests");
-  }
-}
-
-async function tendermintSearchIndexUpdated(): Promise<void> {
-  // Tendermint needs some time before a committed transaction is found in search
-  return sleep(75);
-}
-
-function buildKvTx(k: string, v: string): Uint8Array {
-  return toAscii(`${k}=${v}`);
-}
-
-function randomString(): string {
-  const alphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-  return Array.from({ length: 12 })
-    .map(() => alphabet[Math.floor(Math.random() * alphabet.length)])
-    .join("");
-}
-
-function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expected: ExpectedValues): void {
+function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues): void {
   describe("create", () => {
     it("can auto-discover Tendermint version and communicate", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory());
+      const client = await Tendermint34Client.create(rpcFactory());
       const info = await client.abciInfo();
       expect(info).toBeTruthy();
       client.disconnect();
@@ -52,7 +33,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can connect to Tendermint with known version", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint34Client.create(rpcFactory());
       expect(await client.abciInfo()).toBeTruthy();
       client.disconnect();
     });
@@ -60,7 +41,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can get genesis", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
     const genesis = await client.genesis();
     expect(genesis).toBeTruthy();
     client.disconnect();
@@ -68,7 +49,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can broadcast a transaction", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
     const tx = buildKvTx(randomString(), randomString());
 
     const response = await client.broadcastTxCommit({ tx: tx });
@@ -86,9 +67,9 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("gets the same tx hash from backend as calculated locally", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
     const tx = buildKvTx(randomString(), randomString());
-    const calculatedTxHash = adaptor.hashTx(tx);
+    const calculatedTxHash = adaptor34.hashTx(tx);
 
     const response = await client.broadcastTxCommit({ tx: tx });
     expect(response.hash).toEqual(calculatedTxHash);
@@ -98,7 +79,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can query the state", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
 
     const key = randomString();
     const value = randomString();
@@ -117,7 +98,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can get a commit", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
     const response = await client.commit(4);
 
     expect(response).toBeTruthy();
@@ -132,7 +113,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can get validators", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
     const response = await client.validators({});
 
     expect(response).toBeTruthy();
@@ -150,7 +131,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can get all validators", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
     const response = await client.validatorsAll();
 
     expect(response).toBeTruthy();
@@ -168,7 +149,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
   it("can call a bunch of methods", async () => {
     pendingWithoutTendermint();
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
 
     expect(await client.block()).toBeTruthy();
     expect(await client.genesis()).toBeTruthy();
@@ -180,7 +161,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   describe("status", () => {
     it("works", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint34Client.create(rpcFactory());
 
       const status = await client.status();
 
@@ -210,7 +191,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   describe("blockResults", () => {
     it("works", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint34Client.create(rpcFactory());
 
       const height = 3;
       const results = await client.blockResults(height);
@@ -226,7 +207,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   describe("blockchain", () => {
     it("returns latest in descending order by default", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint34Client.create(rpcFactory());
 
       // Run in parallel to increase chance there is no block between the calls
       const [status, blockchain] = await Promise.all([client.status(), client.blockchain()]);
@@ -243,7 +224,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can limit by maxHeight", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint34Client.create(rpcFactory());
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(undefined, height - 1);
@@ -257,7 +238,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("works with maxHeight in the future", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint34Client.create(rpcFactory());
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(undefined, height + 20);
@@ -271,7 +252,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can limit by minHeight and maxHeight", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint34Client.create(rpcFactory());
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(height - 2, height - 1);
@@ -285,7 +266,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("contains all the info", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint34Client.create(rpcFactory());
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(height - 1, height - 1);
@@ -315,7 +296,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   describe("tx", () => {
     it("can query a tx properly", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint34Client.create(rpcFactory());
 
       const find = randomString();
       const me = randomString();
@@ -372,7 +353,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     beforeAll(async () => {
       if (tendermintEnabled()) {
-        const client = await Client.create(rpcFactory(), adaptor);
+        const client = await Tendermint34Client.create(rpcFactory());
 
         // eslint-disable-next-line no-inner-declarations
         async function sendTx(): Promise<void> {
@@ -398,7 +379,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can paginate over txSearch results", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint34Client.create(rpcFactory());
 
       const query = buildQuery({ tags: [{ key: "app.key", value: key }] });
 
@@ -417,7 +398,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
 
     it("can get all search results in one call", async () => {
       pendingWithoutTendermint();
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint34Client.create(rpcFactory());
 
       const query = buildQuery({ tags: [{ key: "app.key", value: key }] });
 
@@ -434,7 +415,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expecte
   });
 }
 
-function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expected: ExpectedValues): void {
+function websocketTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues): void {
   it("can subscribe to block header events", (done) => {
     pendingWithoutTendermint();
 
@@ -442,7 +423,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
 
     (async () => {
       const events: responses.NewBlockHeaderEvent[] = [];
-      const client = await Client.create(rpcFactory(), adaptor);
+      const client = await Tendermint34Client.create(rpcFactory());
       const stream = client.subscribeNewBlockHeader();
       expect(stream).toBeTruthy();
       const subscription = stream.subscribe({
@@ -500,7 +481,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
     const transactionData2 = buildKvTx(randomString(), randomString());
 
     const events: responses.NewBlockEvent[] = [];
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
     const stream = client.subscribeNewBlock();
     const subscription = stream.subscribe({
       next: (event) => {
@@ -556,7 +537,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
     pendingWithoutTendermint();
 
     const events: responses.TxEvent[] = [];
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
     const stream = client.subscribeTx();
     const subscription = stream.subscribe({
       next: (event) => {
@@ -600,7 +581,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
     const transactionData2 = buildKvTx(randomString(), randomString());
 
     const events: responses.TxEvent[] = [];
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
     const query = buildQuery({ tags: [{ key: "app.creator", value: expected.appCreator }] });
     const stream = client.subscribeTx(query);
     expect(stream).toBeTruthy();
@@ -638,7 +619,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
   it("can unsubscribe and re-subscribe to the same stream", async () => {
     pendingWithoutTendermint();
 
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
     const stream = client.subscribeNewBlockHeader();
 
     const event1 = await firstEvent(stream);
@@ -671,7 +652,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
   it("can subscribe twice", async () => {
     pendingWithoutTendermint();
 
-    const client = await Client.create(rpcFactory(), adaptor);
+    const client = await Tendermint34Client.create(rpcFactory());
     const stream1 = client.subscribeNewBlockHeader();
     const stream2 = client.subscribeNewBlockHeader();
 
@@ -683,48 +664,46 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, expec
   });
 }
 
-for (const { url, version, expected } of tendermintInstances) {
-  describe(`Client ${version}`, () => {
-    it("can connect to a given url", async () => {
-      pendingWithoutTendermint();
+describe("Tendermint34Client", () => {
+  const { url, expected } = tendermintInstances[1];
 
-      // default connection
-      {
-        const client = await Client.connect(url);
-        const info = await client.abciInfo();
-        expect(info).toBeTruthy();
-        client.disconnect();
-      }
+  it("can connect to a given url", async () => {
+    pendingWithoutTendermint();
 
-      // http connection
-      {
-        const client = await Client.connect("http://" + url);
-        const info = await client.abciInfo();
-        expect(info).toBeTruthy();
-        client.disconnect();
-      }
+    // default connection
+    {
+      const client = await Tendermint34Client.connect(url);
+      const info = await client.abciInfo();
+      expect(info).toBeTruthy();
+      client.disconnect();
+    }
 
-      // ws connection
-      {
-        const client = await Client.connect("ws://" + url);
-        const info = await client.abciInfo();
-        expect(info).toBeTruthy();
-        client.disconnect();
-      }
-    });
+    // http connection
+    {
+      const client = await Tendermint34Client.connect("http://" + url);
+      const info = await client.abciInfo();
+      expect(info).toBeTruthy();
+      client.disconnect();
+    }
 
-    describe("With HttpClient", () => {
-      const adaptor = adaptorForVersion(version);
-      defaultTestSuite(() => new HttpClient(url), adaptor, expected);
-    });
-
-    describe("With WebsocketClient", () => {
-      // don't print out WebSocket errors if marked pending
-      const onError = process.env.TENDERMINT_ENABLED ? console.error : () => 0;
-      const factory = (): WebsocketClient => new WebsocketClient(url, onError);
-      const adaptor = adaptorForVersion(version);
-      defaultTestSuite(factory, adaptor, expected);
-      websocketTestSuite(factory, adaptor, expected);
-    });
+    // ws connection
+    {
+      const client = await Tendermint34Client.connect("ws://" + url);
+      const info = await client.abciInfo();
+      expect(info).toBeTruthy();
+      client.disconnect();
+    }
   });
-}
+
+  describe("With HttpClient", () => {
+    defaultTestSuite(() => new HttpClient(url), expected);
+  });
+
+  describe("With WebsocketClient", () => {
+    // don't print out WebSocket errors if marked pending
+    const onError = process.env.TENDERMINT_ENABLED ? console.error : () => 0;
+    const factory = (): WebsocketClient => new WebsocketClient(url, onError);
+    defaultTestSuite(factory, expected);
+    websocketTestSuite(factory, expected);
+  });
+});